Refactor MemberCarousel to use custom hook for carousel logic
