+++ src/features/AboutUs/components/MemberCarousel.tsx
@@ -1,45 +1,23 @@
-import React, { useEffect, useRef, useState } from "react";
+import React from "react";
 import MemberCard from "./MemberCard";
 import type { Member } from "./Section";
 import "../layout/TeamCarouselStyle.css";
+import { useMemberCarousel } from "../hooks/useMemberCarousel";
 
 interface MemberCarouselProps {
   members?: Member[];
 }
 
 const MemberCarousel: React.FC<MemberCarouselProps> = ({ members = [] }) => {
-  const trackRef = useRef<HTMLDivElement>(null);
-  const [currentIndex, setCurrentIndex] = useState(0);
-  const [isPaused, setIsPaused] = useState(false);
-  const waitTime = 4000;
-
-  const goToIndex = (index: number) => {
-    setCurrentIndex(index);
-  };
-
-  const goPrev = () => {
-    setCurrentIndex((prev) => (prev - 1 + members.length) % members.length);
-  };
-
-  const goNext = () => {
-    setCurrentIndex((prev) => (prev + 1) % members.length);
-  };
-
-  useEffect(() => {
-    if (isPaused) return;
-
-    const interval = setInterval(() => {
-      setCurrentIndex((prev) => (prev + 1) % members.length);
-    }, waitTime);
-
-    return () => clearInterval(interval);
-  }, [members.length, isPaused]);
-
-  useEffect(() => {
-    if (trackRef.current) {
-      trackRef.current.style.transform = `translateX(-${currentIndex * 100}%)`;
-    }
-  }, [currentIndex]);
+   const {
+    trackRef,
+    currentIndex,
+    setIsPaused,
+    goToIndex,
+    goPrev,
+    goNext,
+  } = useMemberCarousel(members);
+  
 
   return (
     members.length > 0 && (
